refactor DefaultConditionErrorListener & DefaultSlaThresholdErrorListener
